backend/requirements.txt
Django==5.0.3
djangorestframework==3.14.0
mongoengine==0.28.2
django-rest-framework-mongoengine==3.4.1
django-cors-headers==4.3.1
python-dotenv==1.0.1
Pillow==10.2.0
boto3==1.34.44
gunicorn==21.2.0
whitenoise==6.6.0
pymongo==4.3.3
django-cloudinary-storage==0.3.0